+++ src/app/app-routing.module.ts
@@ -2,20 +2,21 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { ContactUsComponent } from './modules/contact-us/contact-us.component';
-import { ProductsListComponent } from './modules/products/components/products-list/products-list.component';
+import { AdminLoadGuard } from './modules/admin/guards/admin-load.guard';
 
 const routes: Routes = [
-  {
-    path: 'products',
-    component: ProductsListComponent
-  },
   {
     path: 'contact-us',
     component: ContactUsComponent
   },
   {
+    path: 'admin',
+    loadChildren: './modules/admin/admin.module#AdminModule',
+    canLoad: [AdminLoadGuard],
+  },
+  {
     path: '',
-    redirectTo: '/products',
+    redirectTo: '/products-list',
     pathMatch: 'full'
   }
 ];