+++ src/app/modules/cart/components/cart-modal/cart-modal.component.ts
@@ -3,6 +3,8 @@ import {
   Component,
   ChangeDetectionStrategy, DoCheck,
 } from '@angular/core';
+import { Router } from '@angular/router';
+
 import { Product } from '../../../products/models/product.model';
 import { CartService } from '../../../../core/services/cart.service';
 
@@ -16,10 +18,10 @@ export class CartModalComponent implements OnInit, DoCheck {
 
   lastUpdated: number;
   sortingOrder: boolean;
-  filteringOrder: boolean;
   selectedSorting: string;
 
   constructor(
+    private router: Router,
     private cartService: CartService
   ) { }
 
@@ -28,7 +30,6 @@ export class CartModalComponent implements OnInit, DoCheck {
 
   ngDoCheck(): void {
     this.lastUpdated = Date.now();
-    console.log(this.filteringOrder);
   }
 
   getProductsFromCart(): Product[] {
@@ -56,8 +57,7 @@ export class CartModalComponent implements OnInit, DoCheck {
   }
 
   order(): void {
-    alert(`Your order price is: $${this.getCartTotal()}`);
-    this.cartService.cleanCart();
+    this.router.navigate(['/order']);
   }
 
   getSortingOrder(): string {