+++ src/app/modules/products/services/products.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@angular/core';
 import { Product } from '../models/product.model';
 import { Categories } from '../models/categories.enum';
 
-import {Observable, of} from 'rxjs';
+import { Observable, of } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -46,5 +47,27 @@ export class ProductsService {
 
   getProducts(): Observable<Product[]> {
     return of(this.productsList);
+  }
+
+  getProduct(id: number | string): Observable<Product> {
+    return this.getProducts().pipe(
+      map((products: Product[]) => products.find(product => product.id === +id))
+    );
+  }
+
+  updateProduct(product: Product): void {
+    const i = this.productsList.findIndex(p => p.id === product.id);
+
+    if (i > -1) {
+      this.productsList.splice(i, 1, product);
+    }
+  }
+
+  deleteProduct(id: number): void {
+    const i = this.productsList.findIndex(p => p.id === id);
+
+    if (i > -1) {
+      this.productsList.splice(i, 1);
+    }
   }
 }