+++ src/app/shared/components/header/header.component.ts
@@ -36,7 +36,18 @@ export class HeaderComponent implements OnInit, AfterViewInit, AfterViewChecked 
     this.router.navigate(['/']);
   }
 
+  navigateToCart(): void {
+    this.router.navigate(['/cart']);
+  }
+
+  navigateToAdmin(): void {
+    this.router.navigate(['/admin']);
+  }
+
   navigateToContacts(): void {
     this.router.navigate(['/contact-us']);
+  }
+  navigateToProductsList(): void {
+    this.router.navigate(['/products-list']);
   }
 }